+++ R/trajectoryMetrics.R
@@ -1,45 +1,36 @@
-#' Metrics for Ecological Trajectory Analysis
+#' Trajectory metrics
 #' 
-#' Ecological Trajectory Analysis (ETA) is a framework to analyze dynamics of ecosystems described as trajectories in a chosen space of multivariate resemblance (De \enc{Cáceres}{Caceres} et al. 2019).
-#' ETA takes trajectories as objects to be analyzed and compared geometrically. Given input trajectory data, the set of functions that provide ETA metrics are:
+#' Set of functions to estimate metrics describing individual trajectories. Given input trajectory data, the set of functions that provide ETA metrics are:
 #' \itemize{
-#' \item{Function \code{trajectoryDistances} calculates the distance between pairs of trajectories.}
 #' \item{Function \code{trajectoryLengths} calculates lengths of directed segments and total path lengths of trajectories.}
 #' \item{Function \code{trajectoryLengths2D} calculates lengths of directed segments and total path lengths of trajectories from 2D coordinates given as input.} 
 #' \item{Function \code{trajectorySpeeds} calculates speeds of directed segments and total path speed of trajectories.}
 #' \item{Function \code{trajectorySpeeds2D} calculates speeds of directed segments and total path speed of trajectories from 2D coordinates given as input.} 
 #' \item{Function \code{trajectoryAngles} calculates the angle between consecutive pairs of directed segments or between segments of ordered triplets of points.}
 #' \item{Function \code{trajectoryAngles2D} calculates the angle between consecutive pairs of directed segments or between segments of ordered triplets of points.}
-#' \item{Function \code{trajectoryConvergence} performs the Mann-Kendall trend test on the distances between trajectories (symmetric test) or the distance between points of one trajectory to the other.}
 #' \item{Function \code{trajectoryDirectionality} calculates (for each trajectory) a statistic that measures directionality of the whole trajectory.}
 #' \item{Function \code{trajectoryVariability} calculates (for each trajectory) a statistic that measures the variability between the states included in the trajectory.}
+#' \item{Function \code{trajectoryMetrics} evaluates several trajectory metrics at once.}
 #' \item{Function \code{trajectoryWindowMetrics} evaluates several trajectory metrics on subtrajectories defined using moving windows.}
 #' }
 #'  
 #' 
 #' @encoding UTF-8
 #' @name trajectoryMetrics
-#' @aliases trajectoryDistances trajectoryLengths trajectoryLengths2D trajectorySpeeds trajectoryAngles trajectoryAngles2D
-#'          trajectoryConvergence trajectoryDirectionality trajectoryVariability
 #' 
 #' @param x An object of class \code{\link{trajectories}}.
-#' @param distance.type The type of distance index to be calculated (Besse et al. 2016; De Cáceres et al. 2019):
-#'   \itemize{
-#'     \item{\code{Hausdorff}: Hausdorff distance between two trajectories.}
-#'     \item{\code{SPD}: Segment path distance.}
-#'     \item{\code{DSPD}: Directed segment path distance (default).}
-#'   }
-#' @param symmetrization Function used to obtain a symmetric distance, so that DSPD(T1,T2) = DSPD(T2,T1) (e.g., \code{mean} or \code{min}). If \code{symmetrization = NULL} then the symmetrization is not conducted and the output dissimilarity matrix is not symmetric. 
+#' @param relativeToInitial Flag to indicate that lengths or angles should be calculated with respect to initial survey.
+#' @param all Flag to indicate that lengths or angles are desired for all segments or for all triangles (i.e. all pairs of segments) in the trajectory. If FALSE, length or angles are calculated according to relativeToInitial flag.
 #' @param add Flag to indicate that constant values should be added (local transformation) to correct triplets of distance values that do not fulfill the triangle inequality.
 #' 
 #' @details 
-#' Details of calculations are given in De \enc{Cáceres}{Caceres} et al (2019). 
+#' Ecological Trajectory Analysis (ETA) is a framework to analyze dynamics of ecosystems described as trajectories in a chosen space of multivariate resemblance (De \enc{Cáceres}{Caceres} et al. 2019).
+#' ETA takes trajectories as objects to be analyzed and compared geometrically. 
+#' 
 #' The input distance matrix \code{d} should ideally be metric. That is, all subsets of distance triplets should fulfill the triangle inequality (see utility function \code{\link{is.metric}}). 
 #' All ETA functions that require metricity include a parameter '\code{add}', which by default is TRUE, meaning that whenever the triangle inequality is broken the minimum constant required to fulfill it is added to the three distances.
 #' If such local (an hence, inconsistent across triplets) corrections are not desired, users should find another way modify \code{d} to achieve metricity, such as PCoA, metric MDS or non-metric MDS (see vignette 'Introduction to Ecological Trajectory Analysis'). 
 #' If parameter '\code{add}' is set to FALSE and problems of triangle inequality exist, ETA functions may provide missing values in some cases where they should not.
-#' 
-#' The resemblance between trajectories is done by adapting concepts and procedures used for the analysis of trajectories in space (i.e. movement data) (Besse et al. 2016).   
 #' 
 #' Function \code{trajectoryAngles} calculates angles between consecutive segments in degrees. For each pair of segments, the angle between the two is defined on the plane that contains the two segments, and measures the change in direction (in degrees) from one segment to the other. 
 #' Angles are always positive, with zero values indicating segments that are in a straight line, and values equal to 180 degrees for segments that are in opposite directions. If \code{all = TRUE}
@@ -55,7 +46,6 @@
 #' permutations with a directional value equal or larger than the observed.
 #'
 #' @return
-#' Function \code{trajectoryDistances} returns an object of class \code{\link{dist}} containing the distances between trajectories (if \code{symmetrization = NULL} then the object returned is of class \code{matrix}). 
 #' 
 #' Functions \code{trajectoryLengths} and  \code{trajectoryLengths2D} return a data frame with the length of each segment on each trajectory and the total length of all trajectories. 
 #' If \code{relativeToInitial = TRUE} lengths are calculated between the initial survey and all the other surveys.
@@ -67,16 +57,12 @@
 #' 
 #' Function \code{trajectoryAngles2D} returns a data frame with angle values on each trajectory. If \code{betweenSegments=TRUE}, then angles are calculated between trajectory segments, alternatively, If \code{betweenSegments=FALSE}, angles are calculated considering Y axis as the North (0°).
 #' 
-#' Function \code{trajectoryConvergence} returns a list with two elements:
-#' \itemize{
-#'   \item{\code{tau}: A matrix with the statistic (Mann-Kendall's tau) of the convergence/divergence test between trajectories. If \code{symmetric=TRUE} then the matrix is square. Otherwise the statistic of the test of the row trajectory approaching the column trajectory.}
-#'   \item{\code{p.value}: A matrix with the p-value of the convergence/divergence test between trajectories. If \code{symmetric=TRUE} then the matrix is square. Otherwise the p-value indicates the test of the row trajectory approaching the column trajectory.}
-#' }
-#' 
 #' Function \code{trajectoryDirectionality} returns a vector with directionality values (one per trajectory). If \code{nperm} is not missing, the function returns a data frame
 #' with a column of directional values and a column of p-values corresponding to the result of the permutational test.
 #' 
 #' Function \code{trajectoryVariability} returns a vector with total variability values (one per trajectory).
+#' 
+#' Function \code{trajectoryMetrics} returns a data frame where rows are trajectories and columns are different trajectory metrics.
 #' 
 #' Function \code{trajectoryWindowMetrics} returns a data frame where rows are midpoints over trajectories and columns correspond to different trajectory metrics.
 #' 
@@ -84,12 +70,10 @@
 #' @author Anthony Sturbois, Vivarmor nature, Réserve Naturelle nationale de la Baie de Saint-Brieuc
 #' 
 #' @references
-#' Besse, P., Guillouet, B., Loubes, J.-M. & François, R. (2016). Review and perspective for distance based trajectory clustering. IEEE Trans. Intell. Transp. Syst., 17, 3306–3317.
-#' 
 #' De \enc{Cáceres}{Caceres} M, Coll L, Legendre P, Allen RB, Wiser SK, Fortin MJ, Condit R & Hubbell S. (2019). 
 #' Trajectory analysis in community ecology. Ecological Monographs 89, e01350.
 #' 
-#' @seealso \code{\link{segmentDistances}}, \code{\link{trajectoryPlot}}, \code{\link{transformTrajectories}}, \code{\link{trajectoryProjection}}
+#' @seealso \code{\link{trajectoryComparison}}, \code{\link{trajectoryPlot}}, \code{\link{transformTrajectories}}, \code{\link{trajectoryProjection}}
 #' 
 #' @examples 
 #' #Description of sites and surveys
@@ -129,172 +113,9 @@
 #' trajectoryAngles2D(xy, sites, surveys, betweenSegments = TRUE)
 #' trajectoryAngles2D(xy, sites, surveys, betweenSegments = FALSE)
 #' 
-#' #Distances between trajectories
-#' trajectoryDistances(x, distance.type = "Hausdorff")
-#' trajectoryDistances(x, distance.type = "DSPD")
-#'   
+#' #Several metrics at once
+#' trajectoryMetrics(x)  
 #'  
-#' @export
-trajectoryDistances<-function(x, distance.type="DSPD", symmetrization = "mean" , add=TRUE) {
-  if(!inherits(x, "trajectories")) stop("'x' should be of class `trajectories`")
-  distance.type <- match.arg(distance.type, c("DSPD", "SPD", "Hausdorff"))
-  
-  d <- x$d
-  surveys <- x$metadata$surveys
-  # This allows treating fixed date trajectories as sites for plotting purposes
-  if(inherits(x, "fd.trajectories")) {
-    sites <- x$metadata$fdT
-  } else if(inherits(x, "cycles")) {
-    sites <- x$metadata$cycles
-  } else if(inherits(x, "sections")) {
-    sites <- x$metadata$sections
-  } else {
-    sites <- x$metadata$sites
-  }
-  
-  siteIDs <- unique(sites)
-  nsite <- length(siteIDs)
-  nsurveysite<-numeric(nsite)
-  
-  for(i in 1:nsite) nsurveysite[i] = sum(sites==siteIDs[i])
-  if(sum(nsurveysite==1)>0) stop("All sites need to be surveyed at least twice")
-  n <- nrow(as.matrix(d))
-  nseg <- sum(nsurveysite)-nsite
-  
-  #Init output
-  dtraj <- matrix(0, nrow=nsite, ncol = nsite)
-  rownames(dtraj) <- siteIDs
-  colnames(dtraj) <- siteIDs
-  if(distance.type=="DSPD"){
-    lsd <- segmentDistances(x, distance.type="directed-segment", add)
-    dsegmat <- as.matrix(lsd$Dseg)
-    for(i1 in 1:nsite) {
-      for(i2 in 1:nsite) {
-        dt12 = 0
-        for(s1 in 1:(nsurveysite[i1]-1)) {
-          dt12ivec = numeric(0)
-          iseg1 = sum(nsurveysite[1:i1]-1)-(nsurveysite[i1]-1)+s1
-          for(s2 in 1:(nsurveysite[i2]-1)) {
-            iseg2 = sum(nsurveysite[1:i2]-1)-(nsurveysite[i2]-1)+s2
-            dt12ivec = c(dt12ivec, dsegmat[iseg1, iseg2])
-          }
-          dt12 = dt12 + min(dt12ivec)
-        }
-        dt12 = dt12/(nsurveysite[i1]-1) #Average of distances between segments of T1 and trajectory T2
-        dt21 = 0 
-        for(s2 in 1:(nsurveysite[i2]-1)) {
-          dt21ivec = numeric(0)
-          iseg2 = sum(nsurveysite[1:i2]-1)-(nsurveysite[i2]-1)+s2
-          for(s1 in 1:(nsurveysite[i1]-1)) {
-            iseg1 = sum(nsurveysite[1:i1]-1)-(nsurveysite[i1]-1)+s1
-            dt21ivec = c(dt21ivec, dsegmat[iseg1, iseg2])
-          }
-          dt21 = dt21 + min(dt21ivec)
-        }
-        dt21 = dt21/(nsurveysite[i2]-1) #Average of distances between segments of T2 and trajectory T1
-        
-        if(!is.null(symmetrization)) {
-          dtraj[i1,i2] = do.call(symmetrization, list(c(dt12,dt21))) #Symmetrization
-          dtraj[i2,i1] = dtraj[i1,i2]
-        } else {
-          dtraj[i1,i2] = dt12
-          dtraj[i2,i1] = dt21
-        }
-      }
-    }
-    
-  } 
-  else if(distance.type=="SPD") {
-    dmat = as.matrix(d)
-    for(i1 in 1:nsite) {
-      ind_surv1 = which(sites==siteIDs[i1])
-      #Surveys may not be in order
-      if(!is.null(surveys)) ind_surv1 = ind_surv1[order(surveys[sites==siteIDs[i1]])]
-      for(i2 in 1:nsite) {
-        ind_surv2 = which(sites==siteIDs[i2])
-        #Surveys may not be in order
-        if(!is.null(surveys)) ind_surv2 = ind_surv2[order(surveys[sites==siteIDs[i2]])]
-        dt12 = 0
-        for(p1 in 1:nsurveysite[i1]) {
-          dt12ivec = numeric(0)
-          ip1 = ind_surv1[p1]
-          for(s2 in 1:(nsurveysite[i2]-1)) {
-            ipi2 = ind_surv2[s2] #initial point
-            ipe2 = ind_surv2[s2+1] #end point
-            dt12ivec = c(dt12ivec, .distanceToSegmentC(dmat[ipi2,ipe2], dmat[ip1, ipi2], dmat[ip1,ipe2], add)[3])
-          }
-          dt12 = dt12 + min(dt12ivec)
-        }
-        dt12 = dt12/nsurveysite[i1] #Average of distances between points of T1 and trajectory T2
-        dt21 = 0 
-        for(p2 in 1:nsurveysite[i2]) {
-          dt21ivec = numeric(0)
-          ip2 = ind_surv2[p2]
-          for(s1 in 1:(nsurveysite[i1]-1)) {
-            ipi1 = ind_surv1[s1] #initial point
-            ipe1 = ind_surv1[s1+1] #end point
-            dt21ivec = c(dt21ivec, .distanceToSegmentC(dmat[ipi1,ipe1], dmat[ip2, ipi1], dmat[ip2,ipe1], add)[3])
-          }
-          dt21 = dt21 + min(dt21ivec)
-        }
-        dt21 = dt21/nsurveysite[i2] #Average of distances between points of T2 and trajectory T1
-        
-        if(!is.null(symmetrization)) {
-          dtraj[i1,i2] = (dt12+dt21)/2 #Symmetrization
-          dtraj[i2,i1] = dtraj[i1,i2]
-        } else {
-          dtraj[i1,i2] = dt12
-          dtraj[i2,i1] = dt21
-        }
-      }
-    }
-  }
-  else if(distance.type=="Hausdorff") {
-    dmat = as.matrix(d)
-    for(i1 in 1:nsite) {
-      ind_surv1 = which(sites==siteIDs[i1])
-      #Surveys may not be in order
-      if(!is.null(surveys)) ind_surv1 = ind_surv1[order(surveys[sites==siteIDs[i1]])]
-      for(i2 in 1:nsite) {
-        ind_surv2 = which(sites==siteIDs[i2])
-        #Surveys may not be in order
-        if(!is.null(surveys)) ind_surv2 = ind_surv2[order(surveys[sites==siteIDs[i2]])]
-        dt12 = 0
-        dt12vec = numeric(0)
-        for(p1 in 1:nsurveysite[i1]) {
-          ip1 = ind_surv1[p1]
-          for(s2 in 1:(nsurveysite[i2]-1)) {
-            ipi2 = ind_surv2[s2] #initial point
-            ipe2 = ind_surv2[s2+1] #end point
-            dt12vec = c(dt12vec, .distanceToSegmentC(dmat[ipi2,ipe2], dmat[ip1, ipi2], dmat[ip1,ipe2], add)[3])
-          }
-        }
-        dt12 = max(dt12vec) #Maximum of distances between points of T1 and segments of T2
-        dt21 = 0 
-        dt21vec = numeric(0)
-        for(p2 in 1:nsurveysite[i2]) {
-          ip2 = ind_surv2[p2]
-          for(s1 in 1:(nsurveysite[i1]-1)) {
-            ipi1 = ind_surv1[s1] #initial point
-            ipe1 = ind_surv1[s1+1] #end point
-            dt21vec = c(dt21vec, .distanceToSegmentC(dmat[ipi1,ipe1], dmat[ip2, ipi1], dmat[ip2,ipe1], add)[3])
-          }
-        }
-        dt21 = max(dt21vec) #Maximum of distances between points of T2 and segments of T1
-        
-        dtraj[i1,i2] = max(dt12, dt21) #maximum of maximums
-        dtraj[i2,i1] = dtraj[i1,i2]
-      }
-    }
-  } 
-  else stop("Wrong distance type")
-  if(!is.null(symmetrization)) return(as.dist(dtraj))
-  return(dtraj)
-}
-
-#' @rdname trajectoryMetrics
-#' @param relativeToInitial Flag to indicate that lengths or angles should be calculated with respect to initial survey.
-#' @param all Flag to indicate that lengths or angles are desired for all segments or for all triangles (i.e. all pairs of segments) in the trajectory. If FALSE, length or angles are calculated according to relativeToInitial flag.
 #' @export
 trajectoryLengths<-function(x, relativeToInitial = FALSE, all=FALSE) {
   if(!inherits(x, "trajectories")) stop("'x' should be of class `trajectories`")
@@ -331,8 +152,8 @@ trajectoryLengths<-function(x, relativeToInitial = FALSE, all=FALSE) {
   if(!all) {
     lengths = as.data.frame(matrix(NA, nrow=nsite, ncol=maxnsurveys))
     row.names(lengths)<-siteIDs
-    if(relativeToInitial) names(lengths)<-c(paste0("Lt1_t",as.character(2:(maxnsurveys))),"Trajectory")
-    else names(lengths)<-c(paste0("S",as.character(1:(maxnsurveys-1))),"Trajectory")
+    if(relativeToInitial) names(lengths)<-c(paste0("Lt1_t",as.character(2:(maxnsurveys))),"Path")
+    else names(lengths)<-c(paste0("S",as.character(1:(maxnsurveys-1))),"Path")
     for(i1 in 1:nsite) {
       ind_surv1 = which(sites==siteIDs[i1])
       #Surveys may not be in order
@@ -447,7 +268,7 @@ trajectorySpeeds<-function(x) {
     for(s in 1:(nsurveysite[i]-1)) {
       speeds[i,s] <- speeds[i,s]/(times[ind_surv[s+1]] - times[ind_surv[s]])
     }
-    speeds[i, maxnsurveys] <- speeds[i, maxnsurveys]/(times[ind_surv[maxnsurveys]] - times[ind_surv[1]])
+    speeds[i, maxnsurveys] <- speeds[i, maxnsurveys]/(times[ind_surv[length(ind_surv)]] - times[ind_surv[1]])
   }
   return(speeds)
 }
@@ -776,82 +597,6 @@ trajectoryAngles2D<-function(xy,sites,surveys,relativeToInitial=FALSE, betweenSe
 }
 
 
-
-#' @rdname trajectoryMetrics
-#' @param symmetric A logical flag to indicate a symmetric convergence comparison of trajectories.
-#' @export
-trajectoryConvergence<-function(x, symmetric = FALSE, add=TRUE){
-  if(!inherits(x, "trajectories")) stop("'x' should be of class `trajectories`")
-  
-  d <- x$d
-  surveys <- x$metadata$surveys
-  # This allows treating fixed date trajectories as sites for plotting purposes
-  if(inherits(x, "fd.trajectories")) {
-    sites <- x$metadata$fdT
-  } else if(inherits(x, "cycles")) {
-    sites <- x$metadata$cycles
-  } else if(inherits(x, "sections")) {
-    sites <- x$metadata$sections
-  } else {
-    sites <- x$metadata$sites
-  }
-  
-  siteIDs <- unique(sites)
-  nsite <- length(siteIDs)
-  nsurveysite<-numeric(nsite)
-  for(i in 1:nsite) nsurveysite[i] <- sum(sites==siteIDs[i])
-  if(sum(nsurveysite<3)>0) stop("All sites need to be surveyed at least three times")
-  n <- nrow(as.matrix(d))
-
-  #Init output
-  tau <- matrix(NA, nrow=nsite, ncol = nsite)
-  rownames(tau) <- siteIDs
-  colnames(tau) <- siteIDs
-  p.value <- tau
-  dmat <- as.matrix(d)
-  for(i1 in 1:(nsite-1)) {
-    ind_surv1 <- which(sites==siteIDs[i1])
-    #Surveys may not be in order
-    if(!is.null(surveys)) ind_surv1 <- ind_surv1[order(surveys[sites==siteIDs[i1]])]
-    for(i2 in (i1+1):nsite) {
-      ind_surv2 <- which(sites==siteIDs[i2])
-      #Surveys may not be in order
-      if(!is.null(surveys)) ind_surv2 <- ind_surv2[order(surveys[sites==siteIDs[i2]])]
-      if(!symmetric) {
-        trajectory <- ind_surv2
-        target <- ind_surv1
-        trajProj <- trajectoryProjection(d,target, trajectory, add=add)
-        dT <- trajProj$distanceToTrajectory
-        mk.test <- MannKendall(dT)
-        tau[i1,i2] <- mk.test$tau
-        p.value[i1,i2] <- mk.test$sl
-        trajectory <- ind_surv1
-        target <- ind_surv2
-        trajProj <- trajectoryProjection(d,target, trajectory, add=add)
-        dT <- trajProj$distanceToTrajectory
-        mk.test <- MannKendall(dT)
-        tau[i2,i1] <- mk.test$tau
-        p.value[i2,i1] <- mk.test$sl
-      } 
-      else {
-        if(length(ind_surv1)==length(ind_surv2)) {
-          dT <- numeric(length(ind_surv1))
-          for(j in 1:length(ind_surv1)) dT[j] = dmat[ind_surv1[j], ind_surv2[j]]
-          mk.test <- MannKendall(dT)
-          tau[i1,i2] <- mk.test$tau
-          p.value[i1,i2] <- mk.test$sl
-          tau[i2,i1] <- mk.test$tau
-          p.value[i2,i1] <- mk.test$sl
-        } else {
-          warning(paste0("sites ",i1, " and ",i2," do not have the same number of surveys."))
-        }
-      }
-    }
-  }
-  return(list(tau = tau, p.value = p.value))
-}
-
-
 #' @rdname trajectoryMetrics
 #' @param nperm The number of permutations to be used in the directionality test.
 #' @export
@@ -933,7 +678,6 @@ trajectoryDirectionality <- function(x, add=TRUE, nperm = NA) {
   }
 }
 
-
 #' @rdname trajectoryMetrics
 #' @export
 trajectoryVariability<-function(x) {
@@ -977,12 +721,38 @@ trajectoryVariability<-function(x) {
   return(var)
 }
 
-
+#' @rdname trajectoryMetrics
+#' @export
+trajectoryMetrics <- function(x, add = TRUE) {
+  if(!inherits(x, "trajectories")) stop("'x' should be of class `trajectories`")
+  if(inherits(x, "fd.trajectories")) {
+    sites <- x$metadata$fdT
+  } else if(inherits(x, "cycles")) {
+    sites <- x$metadata$cycles
+  } else if(inherits(x, "sections")) {
+    sites <- x$metadata$sections
+  } else {
+    sites <- x$metadata$sites
+  }
+  siteIDs <- unique(sites)
+  df <-  data.frame(trajectory = siteIDs, n = NA, 
+                    length = NA, mean_speed = NA, mean_angle = NA,
+                    directionality = NA, variability = NA)
+  for(i in 1:length(siteIDs)) {
+    df$n[i] <- sum(sites==siteIDs[i])
+  }
+  df$length <- trajectoryLengths(x)$Path
+  df$mean_speed <- trajectorySpeeds(x)$Path
+  df$mean_angle <- trajectoryAngles(x, add = add)$mean
+  df$directionality <- trajectoryDirectionality(x, add = add)
+  df$variability <- trajectoryVariability(x)
+  return(df)
+}
 #' @rdname trajectoryMetrics
 #' @param bandwidth Bandwidth of the moving windows (in units of surveys or times, depending on \code{type})
 #' @param type A string, either "surveys" or "times", indicating how windows are defined.
 #' @export
-trajectoryWindowMetrics <- function(x, bandwidth, type = "surveys") {
+trajectoryWindowMetrics <- function(x, bandwidth, type = "surveys", add = TRUE) {
   if(!inherits(x, "trajectories")) stop("'x' should be of class `trajectories`")
   match.arg(type, c("surveys", "times"))
   if(!is.numeric(bandwidth)) stop("'bandwidth' should be an numeric value")
@@ -994,15 +764,16 @@ trajectoryWindowMetrics <- function(x, bandwidth, type = "surveys") {
     if(bandwidth<=0) stop("'bandwidth' should be larger than 0")
   }
   
-  d <- x$d
-  sites <- x$metadata$sites
   surveys <- x$metadata$surveys
   times <- x$metadata$times
-  # This allows treating fixed date trajectories as sites for plotting purposes
   if(inherits(x, "fd.trajectories")) {
     sites <- x$metadata$fdT
   } else if(inherits(x, "cycles")) {
     sites <- x$metadata$cycles
+  } else if(inherits(x, "sections")) {
+    sites <- x$metadata$sections
+  } else {
+    sites <- x$metadata$sites
   }
   
   siteIDs <- unique(sites)
@@ -1011,7 +782,7 @@ trajectoryWindowMetrics <- function(x, bandwidth, type = "surveys") {
   for(i in 1:nsite) nsurveysite[i] = sum(sites==siteIDs[i])
   if(sum(nsurveysite<2)>0) stop("All sites need to be surveyed at least twice")
   
-  df <-  data.frame(sites = sites, midpoint = surveys, n = NA, 
+  df <-  data.frame(trajectory = sites, midpoint = surveys, n = NA, 
                     length = NA, mean_speed = NA, mean_angle = NA,
                     directionality = NA, variability = NA)
   for(i in 1:length(sites)) {
@@ -1024,10 +795,10 @@ trajectoryWindowMetrics <- function(x, bandwidth, type = "surveys") {
     }
     x_i <- subsetTrajectories(x = x, site_selection = sites[i], survey_selection = surveys_window)
     df$n[i] <- length(surveys_window)
-    df$length[i] <- trajectoryLengths(x_i)$Trajectory
-    df$mean_speed[i] <- trajectorySpeeds(x_i)$Trajectory
-    if(length(surveys_window)>2) df$mean_angle[i] <- trajectoryAngles(x_i)$mean
-    if(length(surveys_window)>2) df$directionality[i] <- trajectoryDirectionality(x_i)
+    df$length[i] <- trajectoryLengths(x_i)$Path
+    df$mean_speed[i] <- trajectorySpeeds(x_i)$Path
+    if(length(surveys_window)>2) df$mean_angle[i] <- trajectoryAngles(x_i, add = add)$mean
+    if(length(surveys_window)>2) df$directionality[i] <- trajectoryDirectionality(x_i, add = add)
     df$variability[i] <- trajectoryVariability(x_i)
   }
   return(df)